create modular project with entities

+++ common/src/main/java/am/hovall/common/repositories/ProductRepository.java
@@ -3,8 +3,6 @@ package am.hovall.common.repositories;
 import am.hovall.common.model.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-
-public interface ProductRepository extends JpaRepository<Product, BigDecimal> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
 }
